Move exceptions into their own modules
