+++ robotools/__init__.py
@@ -1,6 +1,8 @@
 from . import evotools, liquidhandling
 from .evotools import InvalidOperationError, Labwares, Tip, Worklist
+from .evotools.exceptions import InvalidOperationError
 from .liquidhandling import Labware, Trough, VolumeOverflowError, VolumeUnderflowError
+from .liquidhandling.exceptions import *
 from .transform import WellRotator, WellShifter, make_well_array, make_well_index_dict
 from .utils import DilutionPlan, get_trough_wells
